Move precompiles and split InnerEvmContext to its file
