+++ crates/revm/src/context/evm_context.rs
@@ -1,26 +1,23 @@
+use super::inner_evm_context::InnerEvmContext;
 use crate::{
     db::Database,
     interpreter::{
-        analysis::to_analysed, gas, return_ok, Contract, CreateInputs, Gas, InstructionResult,
-        Interpreter, InterpreterResult, MAX_CODE_SIZE,
+        return_ok, CallInputs, Contract, Gas, InstructionResult, Interpreter, InterpreterResult,
     },
-    journaled_state::JournaledState,
-    primitives::{
-        keccak256, Account, Address, AnalysisKind, Bytecode, Bytes, CreateScheme, EVMError, Env,
-        HashSet, Spec,
-        SpecId::{self, *},
-        B256, U256,
-    },
-    FrameOrResult, JournalCheckpoint, CALL_STACK_LIMIT,
+    primitives::{Address, Bytes, EVMError, Env, HashSet, U256},
+    ContextPrecompiles, FrameOrResult, CALL_STACK_LIMIT,
 };
-use core::{fmt, ops::{Deref, DerefMut}};
-use revm_interpreter::{SStoreResult, SelfDestructResult};
+use core::{
+    fmt,
+    ops::{Deref, DerefMut},
+};
 use std::boxed::Box;
 
 //#[derive(Debug)]
 pub struct EvmContext<DB: Database> {
     pub inner: InnerEvmContext<DB>,
-    pub precompiles: (),
+    /// Precompiles that are available for evm.
+    pub precompiles: ContextPrecompiles<DB>,
 }
 
 impl<DB: Database + Clone> Clone for EvmContext<DB>
@@ -30,7 +27,7 @@ where
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
-            precompiles: (),
+            precompiles: ContextPrecompiles::default(),
         }
     }
 }
@@ -66,7 +63,7 @@ impl<DB: Database> EvmContext<DB> {
     pub fn new(db: DB) -> Self {
         Self {
             inner: InnerEvmContext::new(db),
-            precompiles: (),
+            precompiles: ContextPrecompiles::default(),
         }
     }
 
@@ -75,7 +72,7 @@ impl<DB: Database> EvmContext<DB> {
     pub fn new_with_env(db: DB, env: Box<Env>) -> Self {
         Self {
             inner: InnerEvmContext::new_with_env(db, env),
-            precompiles: (),
+            precompiles: ContextPrecompiles::default(),
         }
     }
 
@@ -86,386 +83,324 @@ impl<DB: Database> EvmContext<DB> {
     pub fn with_db<ODB: Database>(self, db: ODB) -> EvmContext<ODB> {
         EvmContext {
             inner: self.inner.with_db(db),
-            precompiles: (),
-        }
-    }
-}
-
-/// EVM contexts contains data that EVM needs for execution.
-#[derive(Debug)]
-pub struct InnerEvmContext<DB: Database> {
-    /// EVM Environment contains all the information about config, block and transaction that
-    /// evm needs.
-    pub env: Box<Env>,
-    /// EVM State with journaling support.
-    pub journaled_state: JournaledState,
-    /// Database to load data from.
-    pub db: DB,
-    /// Error that happened during execution.
-    pub error: Result<(), EVMError<DB::Error>>,
-    /// Used as temporary value holder to store L1 block info.
-    #[cfg(feature = "optimism")]
-    pub l1_block_info: Option<crate::optimism::L1BlockInfo>,
-}
-
-impl<DB: Database + Clone> Clone for InnerEvmContext<DB>
-where
-    DB::Error: Clone,
-{
-    fn clone(&self) -> Self {
-        Self {
-            env: self.env.clone(),
-            journaled_state: self.journaled_state.clone(),
-            db: self.db.clone(),
-            error: self.error.clone(),
-            #[cfg(feature = "optimism")]
-            l1_block_info: self.l1_block_info.clone(),
-        }
-    }
-}
-
-impl<DB: Database> InnerEvmContext<DB> {
-    pub fn new(db: DB) -> Self {
-        Self {
-            env: Box::default(),
-            journaled_state: JournaledState::new(SpecId::LATEST, HashSet::new()),
-            db,
-            error: Ok(()),
-            #[cfg(feature = "optimism")]
-            l1_block_info: None,
+            precompiles: ContextPrecompiles::default(),
         }
     }
 
-    /// Creates a new context with the given environment and database.
+    /// Sets precompiles
     #[inline]
-    pub fn new_with_env(db: DB, env: Box<Env>) -> Self {
-        Self {
-            env,
-            journaled_state: JournaledState::new(SpecId::LATEST, HashSet::new()),
-            db,
-            error: Ok(()),
-            #[cfg(feature = "optimism")]
-            l1_block_info: None,
-        }
+    pub fn set_precompiles(&mut self, precompiles: ContextPrecompiles<DB>) {
+        // set warm loaded addresses.
+        self.journaled_state.warm_preloaded_addresses =
+            precompiles.addresses().copied().collect::<HashSet<_>>();
+        self.precompiles = precompiles;
     }
 
-    /// Sets the database.
-    ///
-    /// Note that this will ignore the previous `error` if set.
+    /// Call precompile contract
     #[inline]
-    pub fn with_db<ODB: Database>(self, db: ODB) -> InnerEvmContext<ODB> {
-        InnerEvmContext {
-            env: self.env,
-            journaled_state: self.journaled_state,
-            db,
-            error: Ok(()),
-            #[cfg(feature = "optimism")]
-            l1_block_info: self.l1_block_info,
-        }
-    }
-
-    /// Returns the configured EVM spec ID.
-    #[inline]
-    pub const fn spec_id(&self) -> SpecId {
-        self.journaled_state.spec
-    }
-
-    /// Load access list for berlin hard fork.
-    ///
-    /// Loading of accounts/storages is needed to make them warm.
-    #[inline]
-    pub fn load_access_list(&mut self) -> Result<(), EVMError<DB::Error>> {
-        for (address, slots) in self.env.tx.access_list.iter() {
-            self.journaled_state
-                .initial_account_load(*address, slots, &mut self.db)?;
-        }
-        Ok(())
-    }
-
-    /// Return environment.
-    #[inline]
-    pub fn env(&mut self) -> &mut Env {
-        &mut self.env
-    }
-
-    /// Fetch block hash from database.
-    #[inline]
-    pub fn block_hash(&mut self, number: U256) -> Result<B256, EVMError<DB::Error>> {
-        self.db.block_hash(number).map_err(EVMError::Database)
-    }
-
-    /// Mark account as touched as only touched accounts will be added to state.
-    #[inline]
-    pub fn touch(&mut self, address: &Address) {
-        self.journaled_state.touch(address);
-    }
-
-    /// Loads an account into memory. Returns `true` if it is cold accessed.
-    #[inline]
-    pub fn load_account(
+    fn call_precompile(
         &mut self,
         address: Address,
-    ) -> Result<(&mut Account, bool), EVMError<DB::Error>> {
-        self.journaled_state.load_account(address, &mut self.db)
-    }
+        input_data: &Bytes,
+        gas: Gas,
+    ) -> Option<InterpreterResult> {
+        let out = self
+            .precompiles
+            .call(address, input_data, gas.limit(), &mut self.inner)?;
 
-    /// Load account from database to JournaledState.
-    ///
-    /// Return boolean pair where first is `is_cold` second bool `exists`.
-    #[inline]
-    pub fn load_account_exist(
-        &mut self,
-        address: Address,
-    ) -> Result<(bool, bool), EVMError<DB::Error>> {
-        self.journaled_state
-            .load_account_exist(address, &mut self.db)
-    }
+        let mut result = InterpreterResult {
+            result: InstructionResult::Return,
+            gas,
+            output: Bytes::new(),
+        };
 
-    /// Return account balance and is_cold flag.
-    #[inline]
-    pub fn balance(&mut self, address: Address) -> Result<(U256, bool), EVMError<DB::Error>> {
-        self.journaled_state
-            .load_account(address, &mut self.db)
-            .map(|(acc, is_cold)| (acc.info.balance, is_cold))
-    }
-
-    /// Return account code and if address is cold loaded.
-    #[inline]
-    pub fn code(&mut self, address: Address) -> Result<(Bytecode, bool), EVMError<DB::Error>> {
-        self.journaled_state
-            .load_code(address, &mut self.db)
-            .map(|(a, is_cold)| (a.info.code.clone().unwrap(), is_cold))
-    }
-
-    /// Get code hash of address.
-    #[inline]
-    pub fn code_hash(&mut self, address: Address) -> Result<(B256, bool), EVMError<DB::Error>> {
-        let (acc, is_cold) = self.journaled_state.load_code(address, &mut self.db)?;
-        if acc.is_empty() {
-            return Ok((B256::ZERO, is_cold));
+        match out {
+            Ok((gas_used, data)) => {
+                if result.gas.record_cost(gas_used) {
+                    result.result = InstructionResult::Return;
+                    result.output = data;
+                } else {
+                    result.result = InstructionResult::PrecompileOOG;
+                }
+            }
+            Err(e) => {
+                result.result = if e == crate::precompile::Error::OutOfGas {
+                    InstructionResult::PrecompileOOG
+                } else {
+                    InstructionResult::PrecompileError
+                };
+            }
         }
-        Ok((acc.info.code_hash, is_cold))
+        Some(result)
     }
 
-    /// Load storage slot, if storage is not present inside the account then it will be loaded from database.
+    /// Make call frame
     #[inline]
-    pub fn sload(
+    pub fn make_call_frame(
         &mut self,
-        address: Address,
-        index: U256,
-    ) -> Result<(U256, bool), EVMError<DB::Error>> {
-        // account is always warm. reference on that statement https://eips.ethereum.org/EIPS/eip-2929 see `Note 2:`
-        self.journaled_state.sload(address, index, &mut self.db)
-    }
-
-    /// Storage change of storage slot, before storing `sload` will be called for that slot.
-    #[inline]
-    pub fn sstore(
-        &mut self,
-        address: Address,
-        index: U256,
-        value: U256,
-    ) -> Result<SStoreResult, EVMError<DB::Error>> {
-        self.journaled_state
-            .sstore(address, index, value, &mut self.db)
-    }
-
-    /// Returns transient storage value.
-    #[inline]
-    pub fn tload(&mut self, address: Address, index: U256) -> U256 {
-        self.journaled_state.tload(address, index)
-    }
-
-    /// Stores transient storage value.
-    #[inline]
-    pub fn tstore(&mut self, address: Address, index: U256, value: U256) {
-        self.journaled_state.tstore(address, index, value)
-    }
-
-    /// Selfdestructs the account.
-    #[inline]
-    pub fn selfdestruct(
-        &mut self,
-        address: Address,
-        target: Address,
-    ) -> Result<SelfDestructResult, EVMError<DB::Error>> {
-        self.journaled_state
-            .selfdestruct(address, target, &mut self.db)
-    }
-
-    /// Make create frame.
-    #[inline]
-    pub fn make_create_frame(
-        &mut self,
-        spec_id: SpecId,
-        inputs: &CreateInputs,
+        inputs: &CallInputs,
     ) -> Result<FrameOrResult, EVMError<DB::Error>> {
-        // Prepare crate.
         let gas = Gas::new(inputs.gas_limit);
 
-        let return_error = |e| {
-            Ok(FrameOrResult::new_create_result(
+        let return_result = |instruction_result: InstructionResult| {
+            Ok(FrameOrResult::new_call_result(
                 InterpreterResult {
-                    result: e,
+                    result: instruction_result,
                     gas,
                     output: Bytes::new(),
                 },
-                None,
+                inputs.return_memory_offset.clone(),
             ))
         };
 
         // Check depth
         if self.journaled_state.depth() > CALL_STACK_LIMIT {
-            return return_error(InstructionResult::CallTooDeep);
+            return return_result(InstructionResult::CallTooDeep);
         }
 
-        // Fetch balance of caller.
-        let (caller_balance, _) = self.balance(inputs.caller)?;
+        let (account, _) = self
+            .inner
+            .journaled_state
+            .load_code(inputs.contract, &mut self.inner.db)?;
+        let code_hash = account.info.code_hash();
+        let bytecode = account.info.code.clone().unwrap_or_default();
 
-        // Check if caller has enough balance to send to the created contract.
-        if caller_balance < inputs.value {
-            return return_error(InstructionResult::OutOfFunds);
+        // Create subroutine checkpoint
+        let checkpoint = self.journaled_state.checkpoint();
+
+        // Touch address. For "EIP-158 State Clear", this will erase empty accounts.
+        if inputs.transfer.value == U256::ZERO {
+            self.load_account(inputs.context.address)?;
+            self.journaled_state.touch(&inputs.context.address);
         }
 
-        // Increase nonce of caller and check if it overflows
-        let old_nonce;
-        if let Some(nonce) = self.journaled_state.inc_nonce(inputs.caller) {
-            old_nonce = nonce - 1;
-        } else {
-            return return_error(InstructionResult::Return);
+        // Transfer value from caller to called account
+        if let Some(result) = self.inner.journaled_state.transfer(
+            &inputs.transfer.source,
+            &inputs.transfer.target,
+            inputs.transfer.value,
+            &mut self.inner.db,
+        )? {
+            self.journaled_state.checkpoint_revert(checkpoint);
+            return return_result(result);
         }
 
-        // Create address
-        let mut init_code_hash = B256::ZERO;
-        let created_address = match inputs.scheme {
-            CreateScheme::Create => inputs.caller.create(old_nonce),
-            CreateScheme::Create2 { salt } => {
-                init_code_hash = keccak256(&inputs.init_code);
-                inputs.caller.create2(salt.to_be_bytes(), init_code_hash)
-            }
-        };
-
-        // Load account so it needs to be marked as warm for access list.
-        self.journaled_state
-            .load_account(created_address, &mut self.db)?;
-
-        // create account, transfer funds and make the journal checkpoint.
-        let checkpoint = match self.journaled_state.create_account_checkpoint(
-            inputs.caller,
-            created_address,
-            inputs.value,
-            spec_id,
-        ) {
-            Ok(checkpoint) => checkpoint,
-            Err(e) => {
-                return return_error(e);
-            }
-        };
-
-        let bytecode = Bytecode::new_raw(inputs.init_code.clone());
-
-        let contract = Box::new(Contract::new(
-            Bytes::new(),
-            bytecode,
-            init_code_hash,
-            created_address,
-            inputs.caller,
-            inputs.value,
-        ));
-
-        Ok(FrameOrResult::new_create_frame(
-            created_address,
-            checkpoint,
-            Interpreter::new(contract, gas.limit(), false),
-        ))
-    }
-
-    /// Handles call return.
-    #[inline]
-    pub fn call_return(
-        &mut self,
-        interpreter_result: &InterpreterResult,
-        journal_checkpoint: JournalCheckpoint,
-    ) {
-        // revert changes or not.
-        if matches!(interpreter_result.result, return_ok!()) {
-            self.journaled_state.checkpoint_commit();
-        } else {
-            self.journaled_state.checkpoint_revert(journal_checkpoint);
-        }
-    }
-
-    /// Handles create return.
-    #[inline]
-    pub fn create_return<SPEC: Spec>(
-        &mut self,
-        interpreter_result: &mut InterpreterResult,
-        address: Address,
-        journal_checkpoint: JournalCheckpoint,
-    ) {
-        // if return is not ok revert and return.
-        if !matches!(interpreter_result.result, return_ok!()) {
-            self.journaled_state.checkpoint_revert(journal_checkpoint);
-            return;
-        }
-        // Host error if present on execution
-        // if ok, check contract creation limit and calculate gas deduction on output len.
-        //
-        // EIP-3541: Reject new contract code starting with the 0xEF byte
-        if SPEC::enabled(LONDON)
-            && !interpreter_result.output.is_empty()
-            && interpreter_result.output.first() == Some(&0xEF)
-        {
-            self.journaled_state.checkpoint_revert(journal_checkpoint);
-            interpreter_result.result = InstructionResult::CreateContractStartingWithEF;
-            return;
-        }
-
-        // EIP-170: Contract code size limit
-        // By default limit is 0x6000 (~25kb)
-        if SPEC::enabled(SPURIOUS_DRAGON)
-            && interpreter_result.output.len()
-                > self
-                    .env
-                    .cfg
-                    .limit_contract_code_size
-                    .unwrap_or(MAX_CODE_SIZE)
-        {
-            self.journaled_state.checkpoint_revert(journal_checkpoint);
-            interpreter_result.result = InstructionResult::CreateContractSizeLimit;
-            return;
-        }
-        let gas_for_code = interpreter_result.output.len() as u64 * gas::CODEDEPOSIT;
-        if !interpreter_result.gas.record_cost(gas_for_code) {
-            // record code deposit gas cost and check if we are out of gas.
-            // EIP-2 point 3: If contract creation does not have enough gas to pay for the
-            // final gas fee for adding the contract code to the state, the contract
-            //  creation fails (i.e. goes out-of-gas) rather than leaving an empty contract.
-            if SPEC::enabled(HOMESTEAD) {
-                self.journaled_state.checkpoint_revert(journal_checkpoint);
-                interpreter_result.result = InstructionResult::OutOfGas;
-                return;
+        if let Some(result) = self.call_precompile(inputs.contract, &inputs.input, gas) {
+            if matches!(result.result, return_ok!()) {
+                self.journaled_state.checkpoint_commit();
             } else {
-                interpreter_result.output = Bytes::new();
+                self.journaled_state.checkpoint_revert(checkpoint);
             }
+            Ok(FrameOrResult::new_call_result(
+                result,
+                inputs.return_memory_offset.clone(),
+            ))
+        } else if !bytecode.is_empty() {
+            let contract = Box::new(Contract::new_with_context(
+                inputs.input.clone(),
+                bytecode,
+                code_hash,
+                &inputs.context,
+            ));
+            // Create interpreter and executes call and push new CallStackFrame.
+            Ok(FrameOrResult::new_call_frame(
+                inputs.return_memory_offset.clone(),
+                checkpoint,
+                Interpreter::new(contract, gas.limit(), inputs.is_static),
+            ))
+        } else {
+            self.journaled_state.checkpoint_commit();
+            return_result(InstructionResult::Stop)
         }
-        // if we have enough gas we can commit changes.
-        self.journaled_state.checkpoint_commit();
+    }
+}
 
-        // Do analysis of bytecode straight away.
-        let bytecode = match self.env.cfg.perf_analyse_created_bytecodes {
-            AnalysisKind::Raw => Bytecode::new_raw(interpreter_result.output.clone()),
-            AnalysisKind::Check => {
-                Bytecode::new_raw(interpreter_result.output.clone()).to_checked()
-            }
-            AnalysisKind::Analyse => {
-                to_analysed(Bytecode::new_raw(interpreter_result.output.clone()))
-            }
+/// Test utilities for the [`EvmContext`].
+#[cfg(any(test, feature = "test-utils"))]
+pub(crate) mod test_utils {
+    use super::*;
+    use crate::{
+        db::{CacheDB, EmptyDB},
+        journaled_state::JournaledState,
+        primitives::{address, Address, Bytes, Env, HashSet, SpecId, B256, U256},
+        InnerEvmContext,
+    };
+    use std::boxed::Box;
+
+    /// Mock caller address.
+    pub const MOCK_CALLER: Address = address!("0000000000000000000000000000000000000000");
+
+    /// Creates `CallInputs` that calls a provided contract address from the mock caller.
+    pub fn create_mock_call_inputs(to: Address) -> CallInputs {
+        CallInputs {
+            contract: to,
+            transfer: revm_interpreter::Transfer {
+                source: MOCK_CALLER,
+                target: to,
+                value: U256::ZERO,
+            },
+            input: Bytes::new(),
+            gas_limit: 0,
+            context: revm_interpreter::CallContext {
+                address: MOCK_CALLER,
+                caller: MOCK_CALLER,
+                code_address: MOCK_CALLER,
+                apparent_value: U256::ZERO,
+                scheme: revm_interpreter::CallScheme::Call,
+            },
+            is_static: false,
+            return_memory_offset: 0..0,
+        }
+    }
+
+    /// Creates an evm context with a cache db backend.
+    /// Additionally loads the mock caller account into the db,
+    /// and sets the balance to the provided U256 value.
+    pub fn create_cache_db_evm_context_with_balance(
+        env: Box<Env>,
+        mut db: CacheDB<EmptyDB>,
+        balance: U256,
+    ) -> EvmContext<CacheDB<EmptyDB>> {
+        db.insert_account_info(
+            test_utils::MOCK_CALLER,
+            crate::primitives::AccountInfo {
+                nonce: 0,
+                balance,
+                code_hash: B256::default(),
+                code: None,
+            },
+        );
+        create_cache_db_evm_context(env, db)
+    }
+
+    /// Creates a cached db evm context.
+    pub fn create_cache_db_evm_context(
+        env: Box<Env>,
+        db: CacheDB<EmptyDB>,
+    ) -> EvmContext<CacheDB<EmptyDB>> {
+        EvmContext {
+            inner: InnerEvmContext {
+                env,
+                journaled_state: JournaledState::new(SpecId::CANCUN, HashSet::new()),
+                db,
+                error: Ok(()),
+                #[cfg(feature = "optimism")]
+                l1_block_info: None,
+            },
+            precompiles: ContextPrecompiles::default(),
+        }
+    }
+
+    /// Returns a new `EvmContext` with an empty journaled state.
+    pub fn create_empty_evm_context(env: Box<Env>, db: EmptyDB) -> EvmContext<EmptyDB> {
+        EvmContext {
+            inner: InnerEvmContext {
+                env,
+                journaled_state: JournaledState::new(SpecId::CANCUN, HashSet::new()),
+                db,
+                error: Ok(()),
+                #[cfg(feature = "optimism")]
+                l1_block_info: None,
+            },
+            precompiles: ContextPrecompiles::default(),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use test_utils::*;
+
+    use crate::{
+        db::{CacheDB, EmptyDB},
+        interpreter::InstructionResult,
+        primitives::{address, Bytecode, Bytes, Env, U256},
+        Frame, FrameOrResult, JournalEntry,
+    };
+    use std::boxed::Box;
+
+    // Tests that the `EVMContext::make_call_frame` function returns an error if the
+    // call stack is too deep.
+    #[test]
+    fn test_make_call_frame_stack_too_deep() {
+        let env = Env::default();
+        let db = EmptyDB::default();
+        let mut context = test_utils::create_empty_evm_context(Box::new(env), db);
+        context.journaled_state.depth = CALL_STACK_LIMIT as usize + 1;
+        let contract = address!("dead10000000000000000000000000000001dead");
+        let call_inputs = test_utils::create_mock_call_inputs(contract);
+        let res = context.make_call_frame(&call_inputs);
+        let Ok(FrameOrResult::Result(err)) = res else {
+            panic!("Expected FrameOrResult::Result");
         };
+        assert_eq!(
+            err.interpreter_result().result,
+            InstructionResult::CallTooDeep
+        );
+    }
 
-        // set code
-        self.journaled_state.set_code(address, bytecode);
+    // Tests that the `EVMContext::make_call_frame` function returns an error if the
+    // transfer fails on the journaled state. It also verifies that the revert was
+    // checkpointed on the journaled state correctly.
+    #[test]
+    fn test_make_call_frame_transfer_revert() {
+        let env = Env::default();
+        let db = EmptyDB::default();
+        let mut evm_context = test_utils::create_empty_evm_context(Box::new(env), db);
+        let contract = address!("dead10000000000000000000000000000001dead");
+        let mut call_inputs = test_utils::create_mock_call_inputs(contract);
+        call_inputs.transfer.value = U256::from(1);
+        let res = evm_context.make_call_frame(&call_inputs);
+        let Ok(FrameOrResult::Result(result)) = res else {
+            panic!("Expected FrameOrResult::Result");
+        };
+        assert_eq!(
+            result.interpreter_result().result,
+            InstructionResult::OutOfFunds
+        );
+        let checkpointed = vec![vec![JournalEntry::AccountLoaded { address: contract }]];
+        assert_eq!(evm_context.journaled_state.journal, checkpointed);
+        assert_eq!(evm_context.journaled_state.depth, 0);
+    }
 
-        interpreter_result.result = InstructionResult::Return;
+    #[test]
+    fn test_make_call_frame_missing_code_context() {
+        let env = Env::default();
+        let cdb = CacheDB::new(EmptyDB::default());
+        let bal = U256::from(3_000_000_000_u128);
+        let mut context = create_cache_db_evm_context_with_balance(Box::new(env), cdb, bal);
+        let contract = address!("dead10000000000000000000000000000001dead");
+        let call_inputs = test_utils::create_mock_call_inputs(contract);
+        let res = context.make_call_frame(&call_inputs);
+        let Ok(FrameOrResult::Result(result)) = res else {
+            panic!("Expected FrameOrResult::Result");
+        };
+        assert_eq!(result.interpreter_result().result, InstructionResult::Stop);
+    }
+
+    #[test]
+    fn test_make_call_frame_succeeds() {
+        let env = Env::default();
+        let mut cdb = CacheDB::new(EmptyDB::default());
+        let bal = U256::from(3_000_000_000_u128);
+        let by = Bytecode::new_raw(Bytes::from(vec![0x60, 0x00, 0x60, 0x00]));
+        let contract = address!("dead10000000000000000000000000000001dead");
+        cdb.insert_account_info(
+            contract,
+            crate::primitives::AccountInfo {
+                nonce: 0,
+                balance: bal,
+                code_hash: by.clone().hash_slow(),
+                code: Some(by),
+            },
+        );
+        let mut evm_context = create_cache_db_evm_context_with_balance(Box::new(env), cdb, bal);
+        let call_inputs = test_utils::create_mock_call_inputs(contract);
+        let res = evm_context.make_call_frame(&call_inputs);
+        let Ok(FrameOrResult::Frame(Frame::Call(call_frame))) = res else {
+            panic!("Expected FrameOrResult::Frame(Frame::Call(..))");
+        };
+        assert_eq!(call_frame.return_memory_range, 0..0,);
     }
 }
