+++ crates/revm/src/test_utils.rs
@@ -1,2 +1,2 @@
 #[doc(hidden)]
-pub use crate::context::test_utils::*;
+pub use crate::context::evm_context::test_utils::*;
